Added in additional routes to rWorkout.js

+++ node/src/controllers/cWorkout.js
@@ -35,7 +35,7 @@ async function DeleteWorkout(req,res,next){
     const {workout_id} = res.locals.bodyData;
     //console.log(workout_id);
     try{
-        const delWorkout = await Workout.deleteOne({_id:workout_id});
+        const delWorkout = await Workout.deleteOne({_id:workout_id, user_id:res.locals.user});
         res.send({deleted:true, count:delWorkout});
     }
     catch(e){
@@ -48,11 +48,41 @@ async function DeleteWorkout(req,res,next){
  */
 async function ListWorkouts(req,res,next){
     try{
-        const found = await Workout.find({}).populate({path: "exercises", populate: {path: "motion sets"}});
+        const found = await Workout.find({}).populate({path: "exercises user_id", populate: {path: "motion sets"}});
         res.send({all: found});
     }
     catch(e){
         res.send({message:e.message});
+    }
+}
+
+/** 
+ * List all workouts for a particular user (this will replace ListWorkouts at somepoint)
+ */
+async function ListMyWorkouts(req,res,next){
+    try{
+        const found = await  Workout.find({user_id:res.locals.user})//.populate({path: "exercises", populate: {path: "motion sets"}});
+        console.log(found);
+        res.send({all:found});
+    }
+    catch(e){
+        next(e.message);
+    }
+}
+
+/**
+ * 
+ */
+async function GetWorkout(req,res,next){
+    const {workout_id} = res.locals.bodyData;
+    try{
+        console.log(res.locals.user);
+        const found = await Workout.findOne({_id:workout_id, user_id:res.locals.user}).populate({path:"exercises", populate: {path: "motion sets"}});
+        console.log(found);
+        res.send({workout:found});
+    }
+    catch(e){
+        next(e.message);
     }
 }
 
@@ -67,7 +97,7 @@ async function AddExercise(req,res,next){
         const newEx = await addedExercise.save();
         console.log(newEx);
         try{
-            const assocWorkout = await Workout.findOneAndUpdate({_id:workout_id},{$push: {"exercises" : newEx._id}});
+            const assocWorkout = await Workout.findOneAndUpdate({_id:workout_id,user_id:res.locals.user},{$push: {"exercises" : newEx._id}});
             res.send({added:true, exercises:assocWorkout.exercises});
         }
         catch(e){
@@ -158,4 +188,5 @@ async function RemoveSet(req, res, next) {
 }
 // maybe a finish exercise function which would flag the workout as completed so that new exercises arent added //
 
-module.exports = {CreateWorkout: CreateWorkout,  DeleteWorkout:DeleteWorkout, ListWorkouts:ListWorkouts, AddExercise:AddExercise, RemoveExercise:RemoveExercise, AddSet:AddSet , RemoveSet: RemoveSet}    
+
+module.exports = {GetWorkout:GetWorkout, ListMyWorkouts:ListMyWorkouts, CreateWorkout: CreateWorkout,  DeleteWorkout:DeleteWorkout, ListWorkouts:ListWorkouts, AddExercise:AddExercise, RemoveExercise:RemoveExercise, AddSet:AddSet , RemoveSet: RemoveSet}
